requirements.txt
prettytable==3.11.0
numpy==2.3.2
torch==2.8.0
flask-cors==6.0.1